Update the pkg-config requires list
Clutter depends on various libraries, some of them backend-specific
like the X11 libraries.

Whenever possible, we should add those requirements to the pkg-config
file. For this reason, we have a variable inside the configure.ac
template file which should be filled with the backend-specific modules
we check for during configure time, and then added to the standard
list of dependencies that we write inside the clutter.pc file.