Move the old ChangeLog into its own file
The old ChangeLog is there to track the project history when it
was in SVN -- also because the commit messages we imported from
Subversion were not that great.

Unfortunately, we need a ChangeLog for autotools to work in
GNU mode; for this reason, we can use a dummy ChangeLog redirecting
to git log.

The ChangeLog will be generated on release from the commit
messages, for users of the tarballs.